refactor: migrate keybindings to DDD structure under src/platform/keybindings
- Create src/platform/keybindings/ domain module with flat structure
- Extract KeyComboImpl and KeybindingImpl classes into separate files
- Move keybinding types, defaults, reserved keys, store, and service
- Update all consumers to import from new location
- Move tests alongside source files
- Update stores/README.md and services/README.md

Amp-Thread-ID: https://ampcode.com/threads/T-019c0394-ae44-72d8-86fc-3ccb014e6b50
