+++ src/components/dialog/content/setting/keybinding/KeyComboDisplay.vue
@@ -13,7 +13,7 @@
 import Tag from 'primevue/tag'
 import { computed } from 'vue'
 
-import type { KeyComboImpl } from '@/stores/keybindingStore'
+import type { KeyComboImpl } from '@/platform/keybindings'
 
 const { keyCombo, isModified = false } = defineProps<{
   keyCombo: KeyComboImpl
