+++ src/platform/keybindings/reserved.ts
