+++ src/services/README.md
@@ -71,7 +71,6 @@ The following table lists ALL services in the system as of 2025-09-01:
 | customerEventsService.ts   | Handles customer event tracking and audit logs            | Analytics  |
 | dialogService.ts           | Provides dialog and modal management                      | UI         |
 | extensionService.ts        | Manages extension registration and lifecycle              | Extensions |
-| keybindingService.ts       | Handles keyboard shortcuts and keybindings                | Input      |
 | litegraphService.ts        | Provides utilities for working with the LiteGraph library | Graph      |
 | load3dService.ts           | Manages 3D model loading and visualization                | 3D         |
 | mediaCacheService.ts       | Manages media file caching with blob storage and cleanup  | Media      |
