+++ src/stores/commandStore.ts
@@ -2,10 +2,9 @@ import { defineStore } from 'pinia'
 import { computed, ref } from 'vue'
 
 import { useErrorHandling } from '@/composables/useErrorHandling'
+import type { KeybindingImpl } from '@/platform/keybindings'
+import { useKeybindingStore } from '@/platform/keybindings'
 import type { ComfyExtension } from '@/types/comfy'
-
-import { useKeybindingStore } from './keybindingStore'
-import type { KeybindingImpl } from './keybindingStore'
 
 export interface ComfyCommand {
   id: string
